# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RTN Manga API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /manga:
    get:
      operationId: listManga
      tags: [manga]
      summary: List all manga
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: genre
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of manga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaListResponse"
    post:
      operationId: createManga
      tags: [manga]
      summary: Create new manga (publisher only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MangaInput"
      responses:
        "201":
          description: Created manga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manga"
...
  /manga/{id}:
    get:
      operationId: getManga
      tags: [manga]
      summary: Get manga by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Manga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manga"
        "404":
          description: Not found
    patch:
      operationId: updateManga
      tags: [manga]
      summary: Update manga (publisher only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MangaUpdate"
      responses:
        "200":
          description: Updated manga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manga"
    delete:
      operationId: deleteManga
      tags: [manga]
      summary: Delete manga (publisher only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /manga/{mangaId}/chapters/{chapterId}/pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: Get pages for a chapter
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
        - name: chapterId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
  /manga/{mangaId}/chapters/{chapterId}/pages/reorder:
    post:
      operationId: reorderPages
      tags: [pages]
      summary: Reorder pages
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
        - name: chapterId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageReorderInput"
      responses:
        "200":
          description: Reordered pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
  /manga/{mangaId}/chapters/{chapterId}/pages/{pageId}:
    delete:
      operationId: deletePage
      tags: [pages]
      summary: Delete a single page
      parameters:
        - name: mangaId
          in: path
          required: true
          schema: { type: integer }
        - name: chapterId
          in: path
          required: true
          schema: { type: integer }
        - name: pageId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /publisher/auth:
    post:
      operationId: publisherAuth
      tags: [publisher]
      summary: Authenticate as publisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherAuthInput"
      responses:
        "200":
          description: Auth result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherAuthResult"
...
components:
  schemas:
...
    Manga:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        genres: { type: array, items: { type: string } }
        chapterCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    MangaInput:
      type: object
      required: [title, type, status]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        coverImage: { type: string }
        type: { type: string }
        status: { type: string }
        genres: { type: array, items: { type: string } }
    MangaUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        coverImage: { type: string }
        type: { type: string }
        status: { type: string }
        genres: { type: array, items: { type: string } }
    MangaListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Manga"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Page:
      type: object
      required: [id, chapterId, pageNumber, imageUrl]
      properties:
        id: { type: integer }
        chapterId: { type: integer }
        pageNumber: { type: integer }
        imageUrl: { type: string }
        createdAt: { type: string }
    PageReorderInput:
      type: object
      required: [pageIds]
      properties:
        pageIds:
          type: array
          items: { type: integer }
    PublisherAuthInput:
      type: object
      required: [code]
      properties:
        code: { type: string }
    PublisherAuthResult:
      type: object
      required: [success, token]
      properties:
        success: { type: boolean }
        token: { type: string }
        message: { type: ["string", "null"] }
...