# lib/api-zod/src/generated/types/mangaUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RTN Manga API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MangaUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  coverImage?: string;
  type?: string;
  status?: string;
  genres?: string[];
}